Rename config/apps/OpenSourceAppConfig.py to follow snake case convention
