+++ embedchain/config/apps/open_source_app_config.py
